requirements.txt
cachetools==5.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
google-auth==2.12.0
google-auth-oauthlib==0.5.3
gspread==5.5.0
idna==3.2
oauthlib==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==4.1.0
python-decouple==3.5
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.7
